横スクロールのpadx修正

+++ ui/widgets/select_csv_file.py
@@ -126,6 +126,10 @@ class SelectCsvFile(ctk.CTkFrame):
     # 利用登録者一覧CSVファイル選択
     def select_user_list_csv(self):
         file_path = open_csv()
+
+        if not file_path:
+            return
+
         self.user_list_csv_path.delete(0, "end")
         self.user_list_csv_path.insert(0, file_path)
         self.focus_set()
@@ -133,6 +137,10 @@ class SelectCsvFile(ctk.CTkFrame):
     # 全送付者リストCSVファイル選択
     def select_address_list_csv(self):
         file_path = open_csv()
+
+        if not file_path:
+            return
+
         self.address_list_csv_path.delete(0, "end")
         self.address_list_csv_path.insert(0, file_path)
         self.focus_set()
